README: fix direnv_layout_dir setup
The static setup of `direnv_layout_dir` in `direnvrc` leads to the
creation of separate layout dirs and nix shell caches for each
subdirectory of a project.
Reason: When `direnvrc` is evaluated, `PWD` has the value of the calling
user's shell, instead of pointing to the project directory (the dir containing `.envrc`).

Fix this by dynamically evaluating `$PWD` in a function.
Dynamic evaluation is also required when `.envrc` includes (via `source_*`) `.envrc`
files from other projects because `$PWD` changes in this case.

Cache the computed layout paths because `shasum` is
fairly expensive (13 ms on my desktop system).