fix(doc): document how to enable flake support on NixOS
Due to NixOS making nix-direnv use the stable version of nix by
default we get this error when using flakes:

direnv: using flake
error: 'print-dev-env' is not a recognised command
[…]
error: 'flake' is not a recognised command

This can be worked around by passing `enableFlakes = true` to the
package.

See also #92.