README: Remove "Shell Integration" section
This is better encapsulated as a recommendation in, say, `direnv`'s readme.
Our repo has arguably too much documentation already
and this is only somewhat related to our domain.

We may want to re-instantiate this as a "TIPS.md" or something,
but I don't feel that this belongs in the top-level rendered README.