Only archive flake inputs excluding `self` when registering flake inputs as gcroots
NOTE: this should probably include a check on whether Nix is new enough
to support lazy trees, but for now that's just WIP.

From my understanding we only need the flake's input path itself because
Nix would copy it into the store anyways when evaluating a flake.
However that's not the case anymore when using the lazy-trees branch (so
neither Nix master nor a Nix release).

This change came from the motivation to not have my flakes in my store
unless I explicitly need it. This is because flakes contain secrets in
my case (these are encrypted via git-crypt, so not leaked in the remote
repo), but they're readable by my user in my git checkout. When copying
this flake into my local store, my local store makes these secrets
available for each user & process on my system which is not nice (even
though it's not the end of the world in my case because it's a
single-user machine). However I'd like to avoid that and now I can by
using lazy-trees, this change is just the last bit to achieve this.