Fix reading nixpkgs version without trailing newline
`read` fails if it encounters EOF before the delim, the `.version` file
does not always have a newline, so `read` fails in that case. If the
env has `set -e` then this will abort. So add a newline to the end
always. Doesn’t seem to matter if there is more than one.