Always set nix-direnv-reload as executable
I am not sure why, but on docker bind mounts using virtiofs and gcsFuse,
`-x` is always true for a file which exists.  This means that, before
this commit was merged, when using nix-direnv in manual mode on a
devcontainer with docker desktop on mac, `nix-direnv-reload` would be
created and then not set as executable, and if they tried to run it
users would see an error like `bad interpreter: Permission denied`.
After this commit is merged, these scenarios work.