Add details for `nix profile` installation method
Previous version of these instructions did not make it very clear that
you need to have nix-direnv _together_ with regular direnv.

As a newcomer, I had to learn this the hard way and spend some time
debugging this. I'm sure this will be very helpful to other people like
me.