drop bash4 from ci
We are no longer able to compile bash4. Ideally we could upgrade to
bash5.0 but I also had issues to compile it. I do not want to drop bash5.0 just now because it
was released in 2019 and probably enough people are still using it.
Until we have a better solution, we no longer test this bash version and
support for this version becomes best effort.
For the most part we are not using many features and the code churn
these days is relative low.